backend/processing/tasks.py
import ray
from llamaparse import parse_file
from pytube import YouTube
from PIL import Image
import openpyxl
import os

# Google API imports (stubs for now)
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build

@ray.remote
def process_files(file_path, file_type, use_image_content=False):
    """
    Uses llamaparse to parse PDF, DOCX, XLSX, PPTX, and image files.
    For images, only parses if use_image_content is True. Otherwise, returns the image path for direct use.
    """
    try:
        if file_type == 'image':
            if use_image_content:
                result = parse_file(file_path)
                return {"type": file_type, "content": result, "parsed": True}
            else:
                # Just return the image path for direct use (e.g., upload to Notion)
                return {"type": file_type, "image_path": file_path, "parsed": False}
        else:
            result = parse_file(file_path)
            return {"type": file_type, "content": result}
    except Exception as e:
        return {"type": file_type, "error": str(e)}

@ray.remote
def process_youtube_videos(youtube_urls):
    """
    Downloads and (optionally) transcribes YouTube videos. (Transcription is a stub here.)
    """
    results = []
    for url in youtube_urls:
        try:
            yt = YouTube(url)
            stream = yt.streams.filter(only_audio=True).first()
            output_path = stream.download()
            # Stub: Replace with actual transcription logic
            transcript = f"Downloaded audio for {url} at {output_path} (transcription not implemented)"
            results.append({"url": url, "transcript": transcript})
            os.remove(output_path)  # Clean up
        except Exception as e:
            results.append({"url": url, "error": str(e)})
    return results

@ray.remote
class TranscriptionWorker:
    def transcribe(self, audio_path, user_id):
        # TODO: Implement transcription logic
        result = ""
        # Store result in DB or return
        return {"user_id": user_id, "transcription": result}
