backend/requirements.txt
ray
langgraph
langchain
langchain-anthropic 
langchain-tavily
langchain_core
requests
llamaparse
pytube
Pillow
openpyxl
fastapi
uvicorn[standard]
python-multipart 